services/ai-core/requirements.txt
fastapi[all]==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
pdfplumber==0.11.0
pytest==8.2.2
firebase-admin>=6.5.0
httpx>=0.27.0
google-genai>=0.3.0
python-multipart>=0.0.9
PyMuPDF>=1.24.0
